backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
PyMySQL==1.1.1
cryptography==45.0.6
pydantic-settings==2.10.1
python-dotenv==1.1.1
email-validator==2.3.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
pytest==8.4.2
httpx==0.28.1
